progress on refactor
so far this is looking good...the implementations appear to work well on
all the base types (Vec<u8>, &[u8], BitVec, BitSlice).  still need to
decide what to do with bytes::Bytes: can either add a wrapper in another
crate and bridge the gap there or perhaps put it behind a feature flag
here?